% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenPair.R
\name{GenPair}
\alias{GenPair}
\title{Generate pre-post data for an adverse event}
\usage{
GenPair(n_pairs, pi, s_node, ns_node)
}
\arguments{
\item{n_pairs}{Integer, number of pairs or number of individuals.}

\item{pi}{Vector, probabilities of marginal pre-post multinomial distribution.}

\item{s_node}{Vector, node vector of true signal.}

\item{ns_node}{Vector, node vector of non-true signal.}
}
\value{
A data frame generated for a pre-post contingency table for occurrence of an adverse event.
}
\description{
This function creates data of pre-post contingency table for occurrence of an adverse event.
}
