% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCount.R
\name{GetCount}
\alias{GetCount}
\title{Get count information from data of pre-post contingency table}
\usage{
GetCount(data, x)
}
\arguments{
\item{data}{Data frame, data for counts of pre-post contingency table for treatment A and B.}

\item{x}{Integer, number corresponding to a specific node.}
}
\value{
A data frame for counts of pre-post contingency table for a node.
}
\description{
This function summarize data of pre-post contingency table to count information.
}
