% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMCMCdist.R
\name{GetMCMCdist}
\alias{GetMCMCdist}
\title{Run Markov Chain Monte Carlo simulation}
\usage{
GetMCMCdist(data, n_iter, n_core = 10)
}
\arguments{
\item{data}{Data frame, data for counts of pre-post contingency table for treatment A and B.}

\item{n_iter}{Integer, number of MCMC iterations.}

\item{n_core}{Integer, number of cores for parallel process.}
}
\value{
A vector for empirical MCMC distribution.
}
\description{
This function run Markov Chain Monte Carlo (MCMC) simulation and get empirical MCMC distribution.
}
