% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetStatistic.R
\name{GetStatistic}
\alias{GetStatistic}
\title{Get conditional background risk TreeScan statistic}
\usage{
GetStatistic(data)
}
\arguments{
\item{data}{Data frame, data for counts of pre-post contingency table for treatment A and B.}
}
\value{
A data frame for counts and statistics of pre-post contingency table for a node.
}
\description{
This function calculate conditional background risk TreeScan statistic from the data.
}
