% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSimulation.R
\name{RunSimulation}
\alias{RunSimulation}
\title{Run simulation analysis}
\usage{
RunSimulation(
  n_AE,
  n_pairs,
  n_signal,
  RR_A,
  RR_B,
  null_pi_A,
  null_pi_B,
  n_MCMC,
  n_cores
)
}
\arguments{
\item{n_AE}{Integer, number of adverse events.}

\item{n_pairs}{Integer, number of pairs or number of individuals.}

\item{RR_A}{Numeric, relative risk for treatment A.}

\item{RR_B}{Numeric, relative risk for treatment B.}

\item{null_pi_A}{Vector, null probabilities of marginal pre-post multinomial distribution for treatment A.}

\item{null_pi_B}{Vector, null probabilities of marginal pre-post multinomial distribution for treatment B.}

\item{n_signals}{Integer, number of true signals.}

\item{n_iter}{Integer, number of MCMC iterations.}

\item{n_core}{Integer, number of cores for parallel process.}
}
\value{
A list containing data frame generated for a specific simulation setting and data frame for true signals.
}
\description{
This function performs a simulation analysis of your simulation settings.
}
