% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTSCBR.R
\name{RunTSCBR}
\alias{RunTSCBR}
\title{Run conditional background risk TreeScan}
\usage{
RunTSCBR(data, n_MCMC, n_core = 10)
}
\arguments{
\item{data}{Data frame, data for counts of pre-post contingency table for treatment A and B.}

\item{n_MCMC}{Integer, number of MCMC iterations.}

\item{n_core}{Integer, number of cores for parallel process.}
}
\value{
A data frame for counts, statistics, and signal detection results of pre-post contingency table for a node.
}
\description{
This function run conditional background risk TreeScan.
}
